Revert "ExpansionPanelList elevation as double (#81813)"
This reverts commit 38840b41a9414607156c33296596fd2f2e02ba0a.

+++ packages/flutter/lib/src/material/expansion_panel.dart
@@ -380,8 +380,14 @@ class ExpansionPanelList extends StatefulWidget {
 
   /// Defines elevation for the [ExpansionPanel] while it's expanded.
   ///
+  /// This uses [kElevationToShadow] to simulate shadows, it does not use
+  /// [Material]'s arbitrary elevation feature.
+  ///
+  /// The following values can be used to define the elevation: 0, 1, 2, 3, 4, 6,
+  /// 8, 9, 12, 16, 24.
+  ///
   /// By default, the value of elevation is 2.
-  final double elevation;
+  final int elevation;
 
   @override
   State<StatefulWidget> createState() => _ExpansionPanelListState();
@@ -554,7 +560,7 @@ class _ExpansionPanelListState extends State<ExpansionPanelList> {
     return MergeableMaterial(
       hasDividers: true,
       dividerColor: widget.dividerColor,
-      elevation: widget.elevation,
+      elevation: widget.elevation.toDouble(),
       children: items,
     );
   }
